Remove profile tab and include profile in home and settings
