+++ lib/src/navigation.dart
@@ -9,7 +9,6 @@ import 'package:lichess_mobile/src/utils/l10n.dart';
 import 'package:lichess_mobile/src/styles/puzzle_icons.dart';
 import 'package:lichess_mobile/src/styles/styles.dart';
 import 'package:lichess_mobile/src/view/home/home_tab_screen.dart';
-import 'package:lichess_mobile/src/view/account/profile_tab_screen.dart';
 import 'package:lichess_mobile/src/view/tools/tools_tab_screen.dart';
 import 'package:lichess_mobile/src/view/watch/watch_tab_screen.dart';
 import 'package:lichess_mobile/src/view/puzzle/puzzle_tab_screen.dart';
@@ -18,8 +17,7 @@ enum BottomTab {
   home(Icons.home),
   puzzles(PuzzleIcons.mix),
   tools(CupertinoIcons.wrench),
-  watch(Icons.live_tv),
-  profile(CupertinoIcons.profile_circled);
+  watch(Icons.live_tv);
 
   const BottomTab(this.icon);
 
@@ -35,8 +33,6 @@ enum BottomTab {
         return strings.tools;
       case BottomTab.watch:
         return strings.watch;
-      case BottomTab.profile:
-        return strings.profile;
     }
   }
 }
@@ -55,8 +51,6 @@ final currentNavigatorKeyProvider = Provider<GlobalKey<NavigatorState>>((ref) {
       return toolsNavigatorKey;
     case BottomTab.watch:
       return watchNavigatorKey;
-    case BottomTab.profile:
-      return profileNavigatorKey;
   }
 });
 
@@ -71,8 +65,6 @@ final currentRootScrollControllerProvider = Provider<ScrollController>((ref) {
       return toolsScrollController;
     case BottomTab.watch:
       return watchScrollController;
-    case BottomTab.profile:
-      return profileScrollController;
   }
 });
 
@@ -80,13 +72,11 @@ final homeNavigatorKey = GlobalKey<NavigatorState>(debugLabel: 'home');
 final puzzlesNavigatorKey = GlobalKey<NavigatorState>(debugLabel: 'puzzles');
 final toolsNavigatorKey = GlobalKey<NavigatorState>(debugLabel: 'tools');
 final watchNavigatorKey = GlobalKey<NavigatorState>(debugLabel: 'watch');
-final profileNavigatorKey = GlobalKey<NavigatorState>(debugLabel: 'profile');
 
 final homeScrollController = ScrollController(debugLabel: 'HomeScroll');
 final puzzlesScrollController = ScrollController(debugLabel: 'PuzzlesScroll');
 final toolsScrollController = ScrollController(debugLabel: 'ToolsScroll');
 final watchScrollController = ScrollController(debugLabel: 'WatchScroll');
-final profileScrollController = ScrollController(debugLabel: 'ProfileScroll');
 
 final RouteObserver<PageRoute<void>> rootNavPageRouteObserver =
     RouteObserver<PageRoute<void>>();
@@ -220,11 +210,6 @@ class BottomNavScaffold extends ConsumerWidget {
           navigatorKey: watchNavigatorKey,
           builder: (context) => const WatchTabScreen(),
         );
-      case 4:
-        return _MaterialTabView(
-          navigatorKey: profileNavigatorKey,
-          builder: (context) => const ProfileTabScreen(),
-        );
       default:
         assert(false, 'Unexpected tab');
         return const SizedBox.shrink();
@@ -257,12 +242,6 @@ class BottomNavScaffold extends ConsumerWidget {
           defaultTitle: context.l10n.watch,
           navigatorKey: watchNavigatorKey,
           builder: (context) => const WatchTabScreen(),
-        );
-      case 4:
-        return CupertinoTabView(
-          defaultTitle: context.l10n.profile,
-          navigatorKey: profileNavigatorKey,
-          builder: (context) => const ProfileTabScreen(),
         );
       default:
         assert(false, 'Unexpected tab');
