+++ lib/src/view/home/home_tab_screen.dart
@@ -17,6 +17,8 @@ import 'package:lichess_mobile/src/utils/navigation.dart';
 import 'package:lichess_mobile/src/widgets/board_preview.dart';
 import 'package:lichess_mobile/src/widgets/feedback.dart';
 import 'package:lichess_mobile/src/widgets/platform.dart';
+import 'package:lichess_mobile/src/widgets/buttons.dart';
+import 'package:lichess_mobile/src/widgets/shimmer.dart';
 import 'package:lichess_mobile/src/model/account/account_repository.dart';
 import 'package:lichess_mobile/src/model/common/speed.dart';
 import 'package:lichess_mobile/src/model/auth/auth_session.dart';
@@ -33,9 +35,12 @@ import 'package:lichess_mobile/src/view/auth/sign_in_widget.dart';
 import 'package:lichess_mobile/src/view/puzzle/puzzle_screen.dart';
 import 'package:lichess_mobile/src/view/user/leaderboard_widget.dart';
 import 'package:lichess_mobile/src/view/user/recent_games.dart';
+import 'package:lichess_mobile/src/view/user/perf_cards.dart';
+import 'package:lichess_mobile/src/view/user/user_activity.dart';
 import 'package:lichess_mobile/src/view/play/play_screen.dart';
 import 'package:lichess_mobile/src/view/game/lobby_game_screen.dart';
 import 'package:lichess_mobile/src/view/game/standalone_game_screen.dart';
+import 'package:lichess_mobile/src/view/settings/settings_screen.dart';
 
 final RouteObserver<PageRoute<void>> homeRouteObserver =
     RouteObserver<PageRoute<void>>();
@@ -124,6 +129,7 @@ class _HomeScreenState extends ConsumerState<HomeTabScreen> with RouteAware {
         ),
         actions: const [
           SignInWidget(),
+          _SettingsButton(),
         ],
       ),
       body: RefreshIndicator(
@@ -157,7 +163,8 @@ class _HomeScreenState extends ConsumerState<HomeTabScreen> with RouteAware {
                   const Text('lichess.org'),
                 ],
               ),
-              trailing: const SignInWidget(),
+              leading: const SignInWidget(),
+              trailing: const _SettingsButton(),
             ),
             CupertinoSliverRefreshControl(
               onRefresh: () => _refreshData(),
@@ -296,25 +303,23 @@ class _HomeBody extends ConsumerWidget {
     return connectivity.when(
       data: (data) {
         if (data.isOnline) {
+          final onlineWidgets = [
+            const _OngoingGamePreview(),
+            const _CreateAGame(),
+            const _DailyPuzzle(),
+            const _PerfCards(),
+            const UserActivityWidget(),
+            const RecentGames(),
+            RatingPrefAware(child: LeaderboardWidget()),
+          ];
+
           return defaultTargetPlatform == TargetPlatform.android
               ? ListView(
                   controller: homeScrollController,
-                  children: [
-                    const _OngoingGamePreview(),
-                    const _CreateAGame(),
-                    const _DailyPuzzle(),
-                    const RecentGames(),
-                    RatingPrefAware(child: LeaderboardWidget()),
-                  ],
+                  children: onlineWidgets,
                 )
               : SliverList(
-                  delegate: SliverChildListDelegate([
-                    const _OngoingGamePreview(),
-                    const _CreateAGame(),
-                    const _DailyPuzzle(),
-                    const RecentGames(),
-                    RatingPrefAware(child: LeaderboardWidget()),
-                  ]),
+                  delegate: SliverChildListDelegate(onlineWidgets),
                 );
         } else {
           return defaultTargetPlatform == TargetPlatform.android
@@ -343,6 +348,21 @@ class _HomeBody extends ConsumerWidget {
             ? child
             : const SliverToBoxAdapter(child: child);
       },
+    );
+  }
+}
+
+class _SettingsButton extends StatelessWidget {
+  const _SettingsButton();
+
+  @override
+  Widget build(BuildContext context) {
+    return SettingsButton(
+      onPressed: () => pushPlatformRoute(
+        context,
+        title: context.l10n.settingsSettings,
+        builder: (_) => const SettingsScreen(),
+      ),
     );
   }
 }
@@ -640,6 +660,50 @@ class _OngoingGamePreview extends ConsumerWidget {
         );
       },
       orElse: () => const SizedBox.shrink(),
+    );
+  }
+}
+
+class _PerfCards extends ConsumerWidget {
+  const _PerfCards();
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final account = ref.watch(accountProvider);
+    return account.when(
+      data: (user) {
+        if (user != null) {
+          return PerfCards(user: user);
+        } else {
+          return const SizedBox.shrink();
+        }
+      },
+      loading: () => Shimmer(
+        child: Padding(
+          padding: Styles.bodySectionPadding,
+          child: SizedBox(
+            height: 106,
+            child: ListView.separated(
+              padding: const EdgeInsets.symmetric(vertical: 3.0),
+              scrollDirection: Axis.horizontal,
+              itemCount: 5,
+              separatorBuilder: (context, index) => const SizedBox(width: 10),
+              itemBuilder: (context, index) => ShimmerLoading(
+                isLoading: true,
+                child: Container(
+                  width: 100,
+                  height: 100,
+                  decoration: BoxDecoration(
+                    color: Colors.black,
+                    borderRadius: BorderRadius.circular(10.0),
+                  ),
+                ),
+              ),
+            ),
+          ),
+        ),
+      ),
+      error: (error, stack) => const SizedBox.shrink(),
     );
   }
 }
